requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
psycopg2-binary==2.9.7
pydantic==2.5.0
PyJWT==2.8.0
email-validator==2.1.0